refactor: isolate recurring expenses
