+++ inserisci.html
@@ -143,7 +143,8 @@
 <script type="module" src="js/navbar.js"></script>
   <script type="module">
     import { requireAuth } from "./js/auth.js";
-    import { addSpesa, addSerieRicorrente, materializeSerie } from "./js/spese.js";
+    import { addSpesa } from "./js/spese.js";
+    import { addSerieRicorrente, materializeSerie } from "./js/ricorrenti.js";
     requireAuth();
 
     // default oggi
